requirements.txt
# requirements.txt
# Short-Term Skills Agent Dependencies

# API Clients
requests==2.31.0
google-generativeai==0.3.1

# Data Processing
python-dotenv==1.0.0

# Optional: If you want to add these features later
# pandas==2.1.4  # For advanced data analysis
# beautifulsoup4==4.12.2  # For HTML parsing
# nltk==3.8.1  # For NLP-based skill extraction
streamlit>=1.28.0